+++ src/cf-atomic-component/src/components/AtomicComponent.js
@@ -11,8 +11,6 @@
    ========================================================================== */
 
 const assign = require( '../utilities/object-assign' ).assign;
-const bind = require( '../utilities/function-bind' ).bind;
-const classList = require( '../utilities/dom-class-list' );
 const Delegate = require( 'dom-delegate' ).Delegate;
 const Events = require( '../mixins/Events' );
 const isFunction = require( '../utilities/type-checkers' ).isFunction;
@@ -42,7 +40,7 @@ function AtomicComponent( element, attributes ) {
 }
 
 // Public instance Methods and properties.
-assign( AtomicComponent.prototype, Events, classList, {
+assign( AtomicComponent.prototype, Events, {
 
   tagName: 'div',
 
@@ -59,7 +57,7 @@ assign( AtomicComponent.prototype, Events, classList, {
     }
 
     this.modifiers.forEach( function( modifier ) {
-      if ( classList.contains( this.element, modifier.ui.base ) ) {
+      if ( this.element.classList.contains( modifier.ui.base ) ) {
         if ( modifier.initialize ) {
           this.initializers.push( modifier.initialize );
           delete modifier.initialize;
@@ -198,7 +196,7 @@ assign( AtomicComponent.prototype, Events, classList, {
         if ( isFunction( this[method] ) ) method = this[method];
         if ( method ) {
           match = key.match( delegateEventSplitter );
-          this.delegate( match[1], match[2], bind( method, this ) );
+          this.delegate( match[1], match[2], method.bind( this ) );
         }
       }
     }
@@ -260,7 +258,7 @@ assign( AtomicComponent.prototype, Events, classList, {
  */
 AtomicComponent.extend = function( attributes ) {
 
-/**
+  /**
  * Function used as constructor in order to establish inheritance
  * chain.
  * @returns {AtomicComponent} An instance.
@@ -275,7 +273,7 @@ AtomicComponent.extend = function( attributes ) {
   assign( child, AtomicComponent );
 
   if ( attributes.hasOwnProperty( 'ui' ) &&
-  attributes.ui.hasOwnProperty( 'base' ) ) {
+       attributes.ui.hasOwnProperty( 'base' ) ) {
     child.selector = attributes.ui.base;
   }
 
@@ -288,15 +286,17 @@ AtomicComponent.extend = function( attributes ) {
 /**
  * Function used to instantiate all instances of the particular
  * atomic component on a page.
+ * @param {HTMLNode} scope - Where to search for components within.
  *
  * @returns {Array} List of AtomicComponent instances.
  */
-AtomicComponent.init = function() {
-  const elements = document.querySelectorAll( this.selector );
+AtomicComponent.init = function( scope ) {
+  const base = scope || document;
+  const elements = base.querySelectorAll( this.selector );
   const components = [];
   let element;
 
-  for ( let i = 0; i < elements.length; ++i ) {
+  for ( let i = 0, len = elements.length; i < len; i++ ) {
     element = elements[i];
     if ( element.hasAttribute( 'data-bound' ) === false ) {
       components.push( new this( element ) );