+++ test/unit-test/src/cf-atomic-component/src/utilities/dom-closest-spec.js
@@ -1,6 +1,4 @@
-const srcPath = require( '../src-path' );
-
-let domClosest;
+import { closest } from '../../../../../../src/cf-atomic-component/src/utilities/dom-closest';
 
 let testBlockA;
 let testBlockB;
@@ -9,20 +7,19 @@ let testBlockD;
 let UNDEFINED;
 
 const HTML_SNIPPET = `
-  <section id="test-block-a">
-    <div id="test-block-b">
-      <div id="test-block-c" >
-        <div id="test-block-d"></div>
-      </div>
+<section id="test-block-a">
+  <div id="test-block-b">
+    <div id="test-block-c" >
+      <div id="test-block-d"></div>
     </div>
-  </section>
+  </div>
+</section>
 `;
 
 describe( 'dom-closest', () => {
 
-  beforeAll( () => {
+  beforeEach( () => {
     document.body.innerHTML = HTML_SNIPPET;
-    domClosest = require( srcPath + '/utilities/dom-closest' ).closest;
     testBlockA = document.getElementById( 'test-block-a' );
     testBlockB = document.getElementById( 'test-block-b' );
     testBlockC = document.getElementById( 'test-block-c' );
@@ -31,51 +28,53 @@ describe( 'dom-closest', () => {
 
   it( 'should find the current DOM node if the node matches the selector',
     () => {
-      let element = domClosest( testBlockD, 'div' );
+      let element = closest( testBlockD, 'div' );
       expect( element === testBlockD ).toBe( true );
-      element = domClosest( testBlockD, 'div div' );
+      element = closest( testBlockD, 'div div' );
       expect( element === testBlockD ).toBe( true );
     }
   );
 
   it( 'should return null if a node isn\'t found', () => {
-    let element = domClosest( testBlockA, '.test-block' );
+    let element = closest( testBlockA, '.test-block' );
     expect( element === null ).toBe( true );
-    element = domClosest( testBlockA, 'div.test' );
+    element = closest( testBlockA, 'div.test' );
     expect( element === null ).toBe( true );
   } );
 
   it( 'should return the correct parent node', () => {
-    let element = domClosest( testBlockD, 'section' );
+    let element = closest( testBlockD, 'section' );
     expect( element === testBlockA ).toBe( true );
-    element = domClosest( testBlockC, 'section > div' );
+    element = closest( testBlockC, 'section > div' );
     expect( element === testBlockB ).toBe( true );
   } );
 
   it( 'should use the native closest method if it exists', () => {
     const spy = testBlockD.closest = jest.fn();
-    domClosest( testBlockD, 'section' );
+    closest( testBlockD, 'section' );
     expect( spy ).toHaveBeenCalled();
   } );
 
   it( 'should use the correct matches method', () => {
     const spy = jest.fn();
-    delete testBlockD.closest;
+    /* Set the prototype to a generic object as deleting the "closest" method
+    with `delete testBlockD.closest` does not work for some reason. */
+    Object.setPrototypeOf( testBlockD, Object );
     testBlockD.matches = UNDEFINED;
     testBlockD.webkitMatchesSelector = spy;
-    domClosest( testBlockD, 'section' );
+    closest( testBlockD, 'section' );
     expect( spy ).toHaveBeenCalled();
 
     testBlockD.webkitMatchesSelector = UNDEFINED;
     spy.mockReset();
     testBlockD.mozMatchesSelector = spy;
-    domClosest( testBlockD, 'section' );
+    closest( testBlockD, 'section' );
     expect( spy ).toHaveBeenCalled();
 
     testBlockD.mozMatchesSelector = UNDEFINED;
     spy.mockReset();
     testBlockD.msMatchesSelector = spy;
-    domClosest( testBlockD, 'section' );
+    closest( testBlockD, 'section' );
     expect( spy ).toHaveBeenCalled();
   } );
 } );